refactor: Card의 일급컬렉션 Cards 생성
